# backend/doc/frontend-api/swagger.yaml
swagger: "2.0"
info:
  description: "Proof of concept of an asset gateway to join OAX network. A user can submit KYC documents and then mint, burn fiat token via this asset gateway."
  version: "1.0.0"
  title: "OAX Asset Gateway POC"
  termsOfService: ""
  contact:
    email: "projectstallman@openanx.org"
  license:
    name: "Apache 2.0"
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:5000"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
...
  /users/sign-in:
    get:
      tags:
      - "user"
      summary: "Sign in existing user"
      operationId: "signInUser"
      produces:
      - "application/vnd.api+json"
      parameters:
      - name: "authorization"
        in: "header"
        description: "username:password hash for Basic Authorization, see https://docs.oracle.com/cd/E24191_01/common/tutorials/authn_http_basic.html for examples"
        required: true
        type: "string"
      - name: "authorization-ethereum-address"
        in: "header"
        description: "The ethereum address of the user for sign in clear text"
        required: true
        type: "string"
      responses:
        200:
          description: "Successfully sign in user"
        400:
          description: "Input fields doesn't pass validation check"
          schema:
            $ref: "#/definitions/Errors"
        401:
          description: "Sign in failed with wrong combination of email+password"
          schema:
            $ref: "#/definitions/Errors"
        403:
          description: "User is blocked from using the site"
          schema:
            $ref: "#/definitions/Errors"
...
definitions:
...
  Errors:
    type: object
    properties:
      errors:
        type: array
        items:
          $ref: "#/definitions/Error"
  Error:
    type: object
    required: ["title"]
    properties:
      id:
        type: string
        example: "000001"
      links:
        type: object
        properties:
          about:
            type: string
            example: "https://oax.org/asset-gateway/api/v1/error/000001"
      status:
        type: string
        example: "400"
      code:
        type: string
        example: 'oax.error.000001'
      title:
        type: string
        example: "error.email.invalid.format"
      detail:
        type: string
        example: 'error.email.invalid.format.detail'
      source:
        type: object
        properties:
          pointer:
            type: string
            example: "/data/attributes/email"
          parameter:
            type: string
            example: "email"
      meta:
        $ref: "#/definitions/Meta"
  Meta:
    type: object
...